task #7: Fix rounding issue
- Fix issue where the execution time was always being rounded to the nearest
  millisecond by casting to double
- Add test to detect total session duration being rounded

+++ PerformanceRecorder/Attribute/PerformanceLoggingAttribute.cs
@@ -36,7 +36,7 @@ namespace PerformanceRecorder.Attribute
 
         private double GetCurrentTimeInMs()
         {
-            return DateTime.Now.Ticks / TimeSpan.TicksPerMillisecond;
+            return (double)DateTime.Now.Ticks / TimeSpan.TicksPerMillisecond;
         }
 
         private IMethodDefinition GenerateMethodDefinition(Type parentType, string methodName)
